+++ SwiftUIDemo/SwiftUIDemo/DemoMessage.swift
@@ -11,12 +11,13 @@ import SwiftMessages
 struct DemoMessage: Identifiable {
     let title: String
     let body: String
+    let style: DemoMessageView.Style
 
     var id: String { title + body }
 }
 
 extension DemoMessage: MessageViewConvertible {
     func asMessageView() -> DemoMessageView {
-        DemoMessageView(message: self)
+        DemoMessageView(message: self, style: style)
     }
 }