+++ SwiftUIDemo/SwiftUIDemo/DemoMessageView.swift
@@ -7,11 +7,20 @@
 
 import SwiftUI
 
+// A card-style message view
 struct DemoMessageView: View {
 
     // MARK: - API
 
+    enum Style {
+        case standard
+        case card
+        case tab
+    }
+
     let message: DemoMessage
+    let style: Style
+
 
     // MARK: - Variables
 
@@ -20,21 +29,42 @@ struct DemoMessageView: View {
     // MARK: - Body
 
     var body: some View {
+        switch style {
+        case .standard:
+            content()
+                // Mask the content and extend background into the safe area.
+                .mask {
+                    Rectangle()
+                        .edgesIgnoringSafeArea(.top)
+                }
+        case .card:
+            content()
+                // Mask the content with a rounded rectangle
+                .mask {
+                    RoundedRectangle(cornerRadius: 15)
+                }
+                // External padding around the card
+                .padding(10)
+        case .tab:
+            content()
+                // Mask the content with rounded bottom edge and extend background into the safe area.
+                .mask {
+                    UnevenRoundedRectangle(bottomLeadingRadius: 15, bottomTrailingRadius: 15)
+                        .edgesIgnoringSafeArea(.top)
+                }
+        }
+    }
+
+    @ViewBuilder private func content() -> some View {
         VStack(alignment: .leading) {
             Text(message.title).font(.system(size: 20, weight: .bold))
             Text(message.body)
         }
         .multilineTextAlignment(.leading)
+        // Internal padding of the card
         .padding(30)
+        // Greedy width
         .frame(maxWidth: .infinity)
         .background(.demoMessageBackground)
-        // This makes a tab-style view where the bottom corners are rounded and the view's background
-        // extends to the top edge.
-        .mask(
-            UnevenRoundedRectangle(
-                cornerRadii: .init(bottomLeading: 15, bottomTrailing: 15)
-            )
-            .edgesIgnoringSafeArea(.top)
-        )
     }
 }