+++ SwiftUIDemo/SwiftUIDemo/DemoView.swift
@@ -13,10 +13,30 @@ struct DemoView: View {
     @State var message: DemoMessage?
 
     var body: some View {
-        Button("Show message") {
-            message = DemoMessage(title: "Demo", body: "This is a sample SwiftUI message! This content should be long enough to wrap.")
+        VStack {
+            Button("Show standard message") {
+                message = DemoMessage(
+                    title: "Demo",
+                    body: "This is a sample SwiftUI card-style message! This content should be long enough to wrap.",
+                    style: .standard
+                )
+            }
+            Button("Show card message") {
+                message = DemoMessage(
+                    title: "Demo",
+                    body: "This is a sample SwiftUI card-style message! This content should be long enough to wrap.",
+                    style: .card
+                )
+            }
+            Button("Show tab message") {
+                message = DemoMessage(
+                    title: "Demo",
+                    body: "This is a sample SwiftUI card-style message! This content should be long enough to wrap.",
+                    style: .tab
+                )
+            }
         }
-        .buttonBorderShape(.roundedRectangle(radius: 15))
+        .buttonStyle(.bordered)
         .swiftMessage(message: $message)
     }
 }